+文章点赞

+++ server/follow/follow.js
@@ -28,29 +28,20 @@ router.get('/followerWork',(req,res)=>{
     })
 })
 
-// 关注
+// 关注/取消关注
 router.get('/follow',(req,res)=>{
     let token = Number(req.headers.token);
     let current = global.users.get(token);
     let friend = req.query.friend;
-    let sqlFollow = `INSERT  INTO friend(userID,friendID) VALUES ("${ current } ","${ friend }")`;
+    let status = req.query.status;
+    let sql = ``;
+    if(status == 'true'){
+        sql = `INSERT  INTO friend(userID,friendID) VALUES ("${ current } ","${ friend }")`;
+    }else{
+        sql = `DELETE FROM friend WHERE userID = "${current}" AND friendID = "${friend}"`;
+    }
     new Promise((resolve)=>{
-        db(sqlFollow,()=>{
-            resolve()
-        })
-    }).then(()=>{
-        res.status(200).send(`ok`);
-    })
-})
-
-// 取消关注
-router.get('/unfollow',(req,res)=>{
-    let token = Number(req.headers.token);
-    let current = global.users.get(token);
-    let friend = req.query.friend;
-    let sqlUnFollow = `DELETE FROM friend WHERE userID = "${current}" AND friendID = "${friend}"`;
-    new Promise((resolve)=>{
-        db(sqlUnFollow,()=>{
+        db(sql,()=>{
             resolve()
         })
     }).then(()=>{
